TEDEFO-1415 refactoring, moving some code to a new XmlWriteService.java, it uses SdkService.
